Refactor admFuncVariableIsValid

+++ adm_program/installation/installation.php
@@ -67,7 +67,7 @@ require_once(SERVER_PATH.'/adm_program/system/function.php');
 // Initialize and check the parameters
 
 define('THEME_PATH', 'layout');
-$getMode = admFuncVariableIsValid($_GET, 'mode', 'numeric', array('defaultValue' => 1));
+$getMode = admFuncVariableIsValid($_GET, 'mode', 'int', array('defaultValue' => 1));
 $message = '';
 
 // default database type is always MySQL and must be set because of old config files
@@ -137,7 +137,7 @@ elseif(file_exists('../../config.php'))
     exit();
 }
 
-if($getMode === 1)  // (Default) Choose language
+if($getMode === 1) // (Default) Choose language
 {
     session_destroy();
