+++ adm_program/modules/registration/registration_assign.php
@@ -16,7 +16,7 @@ require_once('../../system/common.php');
 require_once('../../system/login_valid.php');
 
 // Initialize and check the parameters
-$getNewUserId = admFuncVariableIsValid($_GET, 'new_user_id', 'numeric', array('requireValue' => true));
+$getNewUserId = admFuncVariableIsValid($_GET, 'new_user_id', 'int', array('requireValue' => true));
 
 // nur Webmaster duerfen User zuordnen, ansonsten Seite verlassen
 if(!$gCurrentUser->approveUsers())
